Fixing issues for PR
