+++ Cargo.toml
@@ -22,8 +22,9 @@ crate-type = ["lib", "staticlib", "cdylib"]
 default = ["u2f-server", "native-bindings"]
 native-bindings = []
 u2f-server = ["u2f", "webpki"]
-u2f = ["base64", "byteorder", "ring", "untrusted", "serde", "serde_repr", "serde_derive", "serde_json", "serde_bytes"]
-webauthn-server = [ "u2f-server", "bytes"]
+u2f = ["auth_base", "untrusted",  "serde_repr", ]
+webauthn-server = [ "auth_base", "bytes", "serde_cbor"]
+auth_base = ["base64", "byteorder", "ring", "serde", "serde_derive", "serde_json", "serde_bytes"]
 
 [dependencies]
 sha2 = "0.8.0"
