+++ src/lib.rs
@@ -10,7 +10,7 @@ pub mod oath;
 #[cfg(feature = "u2f")]
 pub mod u2f;
 
-#[cfg(feature = "webauthn")]
+#[cfg(feature = "webauthn-server")]
 pub mod webauthn;
 
 #[cfg(feature = "native-bindings")]
